refactor: organize touch in own crate
